Integrated multi-level command parsing.
I hope this isn't overengineered. You can now supply
commands to `build`, like so:

    liz build -v
    This will turn on verbosity, which is currently broken.

    liz build -s .html
    This will append .html to all generated paths.

Okay, this was ultimately a lot of work to be able to switch
between my local environment and the dev environment with a
command flag. But I think it's nice groundwork for
extensibility.

e.g.
    liz build --output-dir "~/Desktop"

I also added a hidden feature, which I think might be too
unclear. You can define a route with a `name` attribute,
leaving out the `path`, which you'd think is required.

The inital beauty of the project was that you could glue
together a bunch of routes in 15 lines of code. This is 200
lines of code. I'll simplify one day, maybe.